requirements.txt
functions-framework==3.4.0
google-generativeai==0.3.2
PyYAML==6.0.1
